+++ tac/agents/v2/base/behaviours.py
@@ -99,23 +99,23 @@ class FIPABehaviour:
         :return: None
         """
 
-        logger.debug("[{}]: on propose as {}.".format(self.name, dialogue.role))
+        logger.debug("[{}]: on propose as {}.".format(self.name, dialogue.role()))
         proposal = propose.proposals[0]
         transaction_id = generate_transaction_id(self.crypto.public_key, propose.destination, dialogue.dialogue_label, dialogue.is_seller)
         transaction = Transaction.from_proposal(proposal,
                                                 transaction_id,
-                                                is_buyer=not dialogue.is_seller,
+                                                is_buyer=not dialogue.is_seller(),
                                                 counterparty=propose.destination,
                                                 sender=self.crypto.public_key,
                                                 crypto=self.crypto)
         new_msg_id = propose.msg_id + 1
         if self._is_profitable_transaction(transaction, dialogue):
-            logger.debug("[{}]: Accepting propose (as {}).".format(self.name, dialogue.role))
+            logger.debug("[{}]: Accepting propose (as {}).".format(self.name, dialogue.role()))
             self.game_instance.lock_manager.add_lock(transaction, as_seller=dialogue.is_seller)
             self.game_instance.lock_manager.add_pending_acceptances(dialogue, new_msg_id, transaction)
             result = Accept(new_msg_id, propose.dialogue_id, propose.destination, propose.msg_id)
         else:
-            logger.debug("[{}]: Declining propose (as {})".format(self.name, dialogue.role))
+            logger.debug("[{}]: Declining propose (as {})".format(self.name, dialogue.role()))
             result = Decline(new_msg_id, propose.dialogue_id, propose.destination, propose.msg_id)
         return result
 
@@ -142,7 +142,7 @@ class FIPABehaviour:
         result = proposal_delta_score >= 0
         logger.debug("[{}]: is good proposal for {}? {}: tx_id={}, "
                      "delta_score={}, amount={}"
-                     .format(self.name, dialogue.role, result, transaction.transaction_id,
+                     .format(self.name, dialogue.role(), result, transaction.transaction_id,
                              proposal_delta_score, transaction.amount))
         return result
 
@@ -150,8 +150,8 @@ class FIPABehaviour:
         """
         On Decline handler.
 
-        :param decline: the decline object
-        :param dialogue: the dialogue object
+        :param decline: the decline
+        :param dialogue: the dialogue
 
         :return: None
         """
@@ -205,12 +205,12 @@ class FIPABehaviour:
         if self._is_profitable_transaction(transaction, dialogue):
             if self.game_instance.is_world_modeling:
                 self.game_instance.world_state.update_on_accept(transaction)
-            logger.debug("[{}]: Locking the current state (as {}).".format(self.name, dialogue.role))
+            logger.debug("[{}]: Locking the current state (as {}).".format(self.name, dialogue.role()))
             self.game_instance.lock_manager.add_lock(transaction, as_seller=dialogue.is_seller)
             results.append(OutContainer(message=transaction.serialize(), message_id=new_msg_id, dialogue_id=accept.dialogue_id, destination=self.game_instance.controller_pbk))
             results.append(Accept(new_msg_id, accept.dialogue_id, accept.destination, accept.msg_id))
         else:
-            logger.debug("[{}]: Decline the accept (as {}).".format(self.name, dialogue.role))
+            logger.debug("[{}]: Decline the accept (as {}).".format(self.name, dialogue.role()))
             results.append(Decline(new_msg_id, accept.dialogue_id, accept.destination, accept.msg_id))
         return results
