+++ tac/agents/v2/base/game_instance.py
@@ -17,7 +17,6 @@
 #   limitations under the License.
 #
 # ------------------------------------------------------------------------------
-import copy
 import datetime
 from enum import Enum
 from typing import List, Optional, Set
@@ -28,8 +27,9 @@ from oef.schema import Description
 
 from tac.agents.v2.base.dialogues import Dialogues
 from tac.agents.v2.base.lock_manager import LockManager
+from tac.agents.v2.base.strategy import BaselineStrategy
 from tac.platform.game import AgentState, WorldState, GameConfiguration
-from tac.helpers.misc import build_query, get_goods_quantities_description, marginal_utility
+from tac.helpers.misc import build_query, get_goods_quantities_description
 from tac.platform.protocol import GameData
 
 
@@ -302,80 +302,4 @@ class GameInstance:
             proposals.append(proposal)
         if proposals == []:
             proposals.append(candidate_proposals[0])  # TODO remove this
-        return proposals
-
-
-class BaselineStrategy:
-
-    def supplied_good_quantities(current_holdings: List[int]) -> List[int]:
-        """
-        Generates list of quantities which are supplied.
-
-        :param current_holdings: a list of current good holdings
-        :return: a list of quantities
-        """
-        return [quantity - 1 for quantity in current_holdings]
-
-    def supplied_good_pbks(good_pbks: List[str], current_holdings: List[int]) -> Set[str]:
-        """
-        Generates set of good pbks which are supplied.
-
-        :param good_pbks: a list of good pbks
-        :param current_holdings: a list of current good holdings
-        :return: a set of pbks
-        """
-        return {good_pbk for good_pbk, quantity in zip(good_pbks, current_holdings) if quantity > 1}
-
-    def demanded_good_quantities(current_holdings: List[int]) -> List[int]:
-        """
-        Generates list of quantities which are demanded.
-
-        :param current_holdings: a list of current good holdings
-        :return: a list of quantities
-        """
-        return [1 for _ in current_holdings]
-
-    def demanded_good_pbks(good_pbks: List[str], current_holdings: List[int]) -> Set[str]:
-        """
-        Generates set of good pbks which are demanded.
-
-        :param good_pbks: a list of good pbks
-        :param current_holdings: a list of current good holdings
-        :return: a set of pbks
-        """
-        return {good_pbk for good_pbk, quantity in zip(good_pbks, current_holdings)}
-
-    def get_proposals(good_pbks: List[str], current_holdings: List[int], utility_params: List[int], tx_fee: float, is_seller: bool, is_world_modeling: bool, world_state: WorldState) -> List[Description]:
-        """
-        Generates proposals from the seller/buyer.
-
-        :param good_pbks: a list of good pbks
-        :param current_holdings: a list of current good holdings
-        :param utility_params: a list of utility params
-        :param tx_fee: the transaction fee
-        :param is_seller: Boolean indicating the role of the agent
-
-        :return: a list of proposals in Description form
-        """
-        quantities = BaselineStrategy.supplied_good_quantities(current_holdings) if is_seller else BaselineStrategy.demanded_good_quantities(current_holdings)
-        proposals = []
-        zeroslist = [0] * len(quantities)
-        rounding_adjustment = 0.01
-        for good_id, good_pbk in zip(range(len(quantities)), good_pbks):
-            if is_seller and quantities[good_id] == 0: continue
-            lis = copy.deepcopy(zeroslist)
-            lis[good_id] = 1
-            desc = get_goods_quantities_description(good_pbks, lis, is_supply=is_seller)
-            delta_holdings = [i * -1 for i in lis] if is_seller else lis
-            switch = -1 if is_seller else 1
-            marginal_utility_from_single_good = marginal_utility(utility_params, current_holdings, delta_holdings) * switch
-            share_of_tx_fee = round(tx_fee / 2.0, 2)
-            if is_world_modeling:
-                desc.values["price"] = world_state.expected_price(good_pbk, round(marginal_utility_from_single_good, 2), is_seller, share_of_tx_fee)
-            else:
-                if is_seller:
-                    desc.values["price"] = round(marginal_utility_from_single_good, 2) + share_of_tx_fee + rounding_adjustment
-                else:
-                    desc.values["price"] = round(marginal_utility_from_single_good, 2) - share_of_tx_fee - rounding_adjustment
-            proposals.append(desc)
         return proposals
